requirements.txt
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
pycparser==2.22
pytesseract==0.3.10
